requirements.txt
google-adk>=1.15.0
google-generativeai>=0.8.3
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
